+++ src/main/javascript/PostMessageAPI/index.js
@@ -1,3 +1,2 @@
-export StateAPIClient from './StateAPIClient';
 export UserAPIClient from './UserAPIClient';