+++ src/main/javascript/index.js
@@ -2,7 +2,10 @@
 export createApp from './Core/createApp';
 export App from './Core/App';
 export * as AppEvents from './Core/AppEvents';
-export * as ContextEvents from './Core/ContextEvents';
+
 export * as UIEvents from './UI/UIEvents';
+export * as StateEvents from './State/StateEvents';
+
+export * as ContextEvents from './Core/ContextEvents';
 export * as TicketEvents from './Context/TicketEvents';